added sender service and prettier
