+++ package.json
@@ -11,6 +11,7 @@
   "license": "ISC",
   "dependencies": {
     "dotenv": "^16.0.3",
-    "nodemailer": "^6.9.1"
+    "nodemailer": "^6.9.1",
+    "prettier": "^2.8.4"
   }
 }
