+++ src/Common/config.js
@@ -1,0 +1,5 @@
+import dotenv from "dotenv"
+
+dotenv.config()
+
+export const { SENDER_EMAIL, SENDER_PASS } = process.env
